feat(): finish task 2, add mapStruct, add spring security, add schedulers

+++ src/main/java/betr/intern/chainsaw/service/UserService.java
@@ -4,6 +4,7 @@ import betr.intern.chainsaw.model.User;
 import betr.intern.chainsaw.repository.UserRepository;
 import java.util.List;
 import java.util.Objects;
+import java.util.Set;
 import java.util.UUID;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,16 +21,16 @@ public class UserService {
         return userRepository.findById(id).orElseGet(() -> null);
     }
 
-    public List<User> findAll() {
-        return userRepository.findAll();
-    }
-
     public User findByEmail(final String email) {
         return userRepository.findByEmail(email);
     }
 
-    public User findByName(final String name) {
-        return userRepository.findByName(name);
+    public List<User> findAll() {
+        return userRepository.findAll();
+    }
+
+    public List<User> findAllById(final Set<UUID> ids) {
+        return userRepository.findAllById(ids);
     }
 
     @Transactional
@@ -49,7 +50,11 @@ public class UserService {
     }
 
     @Transactional
-    public void deleteById(final UUID id) {
+    public String deleteById(final UUID id) {
+        if (userRepository.existsById(id)) {
+            return String.format("User with id=%s did not exist to begin with", id);
+        }
         userRepository.deleteById(id);
+        return String.format("User with id=%s deleted", id);
     }
 }
